Add Color type
I had troubles implementing the diffuse reflection algorithm using the
color.Color interface as it hinders color manipulation.

The color.Color interface was getting in the way of performing color
operations when computing the color of a pixel, so replaced it in favor
of my own color type, that also implements the color.Color interface in
order to be used by objects that implement the image.Image.