Fix bug in Ray.IntersectSphere
A vector operation was inverted. It explains why I had to multiply the Z
and Y components of the spheres by -1 to make them appear in the screen
with the right orientation.

The problem was only noticed after I implemented the diffuse reflection
calculations as the spheres were rendered dark with no shadow at all.